requirement.txt
Flask==2.0.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
itsdangerous==2.0.1
python-dotenv==0.19.2
